refactor BreakpointChanger to use TypesetStyle, create TypesetStyle
