+++ packages/addons-website/scss/components/_website-typeset-example.scss
@@ -25,6 +25,7 @@
 
 .#{$prefix}--typeset-example-description {
   padding: 1rem 25% 1rem 1rem;
+  overflow: hidden;
 }
 
 .#{$prefix}--typeset-example-specs {
