+++ packages/addons-website/src/components/TypesetStyle/TypesetStyle-story.js
@@ -1,0 +1,35 @@
+import { storiesOf } from '@storybook/react';
+import React from 'react';
+import TypesetStyle from '.';
+
+storiesOf('Typeset Style', module)
+.add(
+  'Typeset Style - Productive',
+  () => (
+    <div style={{ width: '100%' }}>
+      <TypesetStyle breakpointControls={false} 
+                    title="Productive" 
+                    typesets="caption,label,helperText,code,body,heading,productHeading" />
+    </div>
+  ),
+  {
+    info: {
+      text: 'Typeset Style - Productive',
+    },
+  }
+)
+.add(
+  'Typeset Style - Expressive',
+  () => (
+    <div style={{ width: '100%' }}>
+      <TypesetStyle breakpointControls={true} 
+                    title="Expressive"
+                    typesets="fluidHeading,fluidParagraph,fluidQuotation,fluidDisplay" />
+    </div>
+  ),
+  {
+    info: {
+      text: 'Typeset Style - Expressive',
+    },
+  }
+);
