+++ packages/addons-website/src/components/TypesetStyle/index.js
@@ -1,0 +1,2 @@
+import TypesetStyle from './TypesetStyle';
+export default TypesetStyle;
